Move GH db parser to freezer.py
